backend/requirements.txt
python-dotenv==0.10.1
SQLAlchemy==1.3.1
Flask==1.0.2
Flask-Cors==3.0.7
waitress==1.3.0
